requirements.txt
alabaster==0.7.13
asgiref==3.8.1
async-timeout==4.0.3
babel==2.16.0
backports.zoneinfo==0.2.1
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
cryptography==2.8
deprecation==2.1.0
diff-match-patch==20241021
Django==4.2.16
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-import-export==4.1.1
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.2
docutils==0.20.1
freezegun==1.5.1
idna==2.8
imagesize==1.4.1
importlib-metadata==8.5.0
jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.1.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.22
pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==19.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==5.3.1
redis==5.1.0
requests==2.32.3
rq==1.16.2
six==1.16.0
snowballstemmer==2.2.0
sphinx==7.1.2
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.5.1
tablib==3.5.0
typing-extensions==4.12.2
urllib3==1.25.8
validators==0.34.0
zipp==3.20.2
